ok show specific task
