+++ README.md
@@ -4,3 +4,7 @@ AIXÔ ES La branca dev temas de David Sergi i Alex
 
 dependencies:
 UUID, colors, inquirer
+
+Para ejecutar programa:
+Desde carpeta /nodeinitialDemo
+npm start
